+++ src/Entities/GatewayAccount.php
@@ -383,13 +383,57 @@ final class GatewayAccount extends Entity
         return Condition::createFromData($data);
     }
 
+    /**
+     * @return string|null
+     */
     public function getAdditionalFilters(): ?string
     {
         return $this->getAttribute('additionalFilters');
     }
 
+    /**
+     * @param string|null $additionalFilters
+     *
+     * @return $this
+     */
     public function setAdditionalFilters(?string $additionalFilters): self
     {
         return $this->setAttribute('additionalFilters', $additionalFilters);
+    }
+
+    /**
+     * @return bool
+     */
+    public function getReconciliationWindowEnabled()
+    {
+        return $this->getAttribute('reconciliationWindowEnabled');
+    }
+
+    /**
+     * @param bool $value
+     *
+     * @return $this
+     */
+    public function setReconciliationWindowEnabled($value)
+    {
+        return $this->setAttribute('reconciliationWindowEnabled', $value);
+    }
+
+    /**
+     * @return int
+     */
+    public function getReconciliationWindowTtl()
+    {
+        return $this->getAttribute('reconciliationWindowTtl');
+    }
+
+    /**
+     * @param int $value
+     *
+     * @return $this
+     */
+    public function setReconciliationWindowTtl($value)
+    {
+        return $this->setAttribute('reconciliationWindowTtl', $value);
     }
 }