+++ tests/Entities/GatewayAccountTest.php
@@ -28,6 +28,8 @@ final class GatewayAccountTest extends TestCase
         $value->setAcquirerName($data['acquirerName']);
         $value->setAdditionalCriteria($data['additionalCriteria']);
         $value->setAdditionalFilters($data['additionalFilters']);
+        $value->setReconciliationWindowEnabled($data['reconciliationWindowEnabled']);
+        $value->setReconciliationWindowTtl($data['reconciliationWindowTtl']);
 
         $expectedJson = $data;
         // Unset read-only properties which we not set.
@@ -52,6 +54,8 @@ final class GatewayAccountTest extends TestCase
         self::assertSame($data['acquirerName'], $gatewayAccount->getAcquirerName());
         self::assertSame($data['additionalCriteria'], $gatewayAccount->getAdditionalCriteria()->jsonSerialize());
         self::assertSame($data['additionalFilters'], $gatewayAccount->getAdditionalFilters());
+        self::assertSame($data['reconciliationWindowEnabled'], $gatewayAccount->getReconciliationWindowEnabled());
+        self::assertSame($data['reconciliationWindowTtl'], $gatewayAccount->getReconciliationWindowTtl());
     }
 
     private static function getDefaultData(): array
@@ -66,6 +70,8 @@ final class GatewayAccountTest extends TestCase
                 'value' => 'website-1',
             ],
             'additionalFilters' => 'websiteId:website-1',
+            'reconciliationWindowEnabled' => true,
+            'reconciliationWindowTtl' => 3600,
         ];
     }
 }