+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Security - in case of vulnerabilities.
 ## [Unreleased]
 
 ### Added
+- [x] Added new properties to `GatewayAccount`: `reconciliationWindowEnabled`, `reconciliationWindowTtl`
+- [x] Added Gateway Account Limit API
+- [x] Added `GatewayAccountLimit`
 - [x] Added new AML API
 - [x] Added new properties to `Transaction`: `planIds`, `isDisputed`, `retriedTransactionId`
 - [x] Added new property to `PaymentCard`: `fingerprint`